+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+* Switched Unity Mediation SDK to version 0.4.0
+* Added Server-to-server (S2S) redeem callback support for rewarded ads
+
 ## 0.0.3
 
 * Fixed build problem for iOS